+++ common/persistence/sql/sqlplugin/postgresql/shard.go
@@ -67,15 +67,15 @@ func (pdb *db) SelectFromShards(filter *sqlplugin.ShardsFilter) (*sqlplugin.Shar
 }
 
 // ReadLockShards acquires a read lock on a single row in shards table
-func (pdb *db) ReadLockShards(filter *sqlplugin.ShardsFilter) (int, error) {
-	var rangeID int
+func (pdb *db) ReadLockShards(filter *sqlplugin.ShardsFilter) (int64, error) {
+	var rangeID int64
 	err := pdb.conn.Get(&rangeID, readLockShardQry, filter.ShardID)
 	return rangeID, err
 }
 
 // WriteLockShards acquires a write lock on a single row in shards table
-func (pdb *db) WriteLockShards(filter *sqlplugin.ShardsFilter) (int, error) {
-	var rangeID int
+func (pdb *db) WriteLockShards(filter *sqlplugin.ShardsFilter) (int64, error) {
+	var rangeID int64
 	err := pdb.conn.Get(&rangeID, lockShardQry, filter.ShardID)
 	return rangeID, err
 }
